+++ lib/fastlane/plugin/wpmreleasetoolkit/actions/common/upload_build_to_apps_cdn.rb
@@ -4,6 +4,7 @@ require 'fastlane/action'
 require 'net/http'
 require 'uri'
 require 'json'
+require_relative '../../helper/apps_cdn_helper'
 
 module Fastlane
   module Actions
@@ -17,8 +18,6 @@ module Fastlane
     class UploadBuildToAppsCdnAction < Action
       # See https://github.a8c.com/Automattic/wpcom/blob/trunk/wp-content/lib/a8c/cdn/src/enums/enum-resource-type.php
       RESOURCE_TYPE = 'Build'
-      # These are from the WordPress.com API, not the Apps CDN plugin
-      VALID_POST_STATUS = %w[publish draft].freeze
       # See https://github.a8c.com/Automattic/wpcom/blob/trunk/wp-content/lib/a8c/cdn/src/enums/enum-build-type.php
       VALID_BUILD_TYPES = %w[
         Alpha
@@ -48,17 +47,13 @@ module Fastlane
         'Full Install',
         'Update',
       ].freeze
-      # See https://github.a8c.com/Automattic/wpcom/blob/trunk/wp-content/lib/a8c/cdn/src/enums/enum-visibility.php
-      VALID_VISIBILITIES = %i[internal external].freeze
-
       def self.run(params)
         UI.message('Uploading build to Apps CDN...')
 
         file_path = params[:file_path]
         UI.user_error!("File not found at path '#{file_path}'") unless File.exist?(file_path)
 
-        api_endpoint = "https://public-api.wordpress.com/rest/v1.1/sites/#{params[:site_id]}/media/new"
-        uri = URI.parse(api_endpoint)
+        uri = Helper::AppsCdnHelper.rest_v1_1_url(site_id: params[:site_id], path: 'media/new')
 
         # Create the request body and headers
         parameters = {
@@ -261,9 +256,7 @@ module Fastlane
             description: 'The visibility of the build (:internal or :external)',
             optional: false,
             type: Symbol,
-            verify_block: proc do |value|
-              UI.user_error!("Visibility must be one of: #{VALID_VISIBILITIES.map { "`:#{_1}`" }.join(', ')}") unless VALID_VISIBILITIES.include?(value.to_s.downcase.to_sym)
-            end
+            verify_block: Helper::AppsCdnHelper.verify_visibility_param
           ),
           FastlaneCore::ConfigItem.new(
             key: :post_status,
@@ -271,9 +264,7 @@ module Fastlane
             optional: true,
             default_value: 'publish',
             type: String,
-            verify_block: proc do |value|
-              UI.user_error!("Post status must be one of: #{VALID_POST_STATUS.join(', ')}") unless VALID_POST_STATUS.include?(value)
-            end
+            verify_block: Helper::AppsCdnHelper.verify_post_status_param
           ),
           FastlaneCore::ConfigItem.new(
             key: :version,